requirements.txt
aiogram==3.10.0
aiosqlite==0.20.0
aiohttp==3.9.5
Pillow==10.3.0
python-dotenv==1.0.1
